requirements.txt
numpy<2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
sentence-transformers==4.1.0
scikit-learn==1.6.1
elasticsearch==9.0.1
torch==2.2.0
google-genai==1.14.0
dotenv==0.9.9
pyngrok==4.3.1
fastapi==0.115.4
uvicorn==0.28.0
nest-asyncio==1.6.0
